Moving _LazyProperty and _lazy_property_deco into gcloud._helpers.
