+++ gcloud/_helpers.py
@@ -59,3 +59,46 @@ class _LocalStack(Local):
         """
         if len(self._stack) > 0:
             return self._stack[-1]
+
+
+class _LazyProperty(object):
+    """Descriptor for lazy loaded property.
+
+    This follows the reify pattern: lazy evaluation and then replacement
+    after evaluation.
+
+    :type name: string
+    :param name: The name of the attribute / property being evaluated.
+
+    :type deferred_callable: callable that takes no arguments
+    :param deferred_callable: The function / method used to evaluate the
+                              property.
+    """
+
+    def __init__(self, name, deferred_callable):
+        self._name = name
+        self._deferred_callable = deferred_callable
+
+    def __get__(self, obj, objtype):
+        if obj is None:
+            return self
+
+        setattr(obj, self._name, self._deferred_callable())
+        return getattr(obj, self._name)
+
+
+def _lazy_property_deco(deferred_callable):
+    """Decorator a method to create a :class:`_LazyProperty`.
+
+    :type deferred_callable: callable that takes no arguments
+    :param deferred_callable: The function / method used to evaluate the
+                              property.
+
+    :rtype: :class:`_LazyProperty`.
+    :returns: A lazy property which defers the deferred_callable.
+    """
+    if isinstance(deferred_callable, staticmethod):
+        # H/T: http://stackoverflow.com/a/9527450/1068170
+        #      For Python2.7+ deferred_callable.__func__ would suffice.
+        deferred_callable = deferred_callable.__get__(True)
+    return _LazyProperty(deferred_callable.__name__, deferred_callable)
